Transfer over dom updates to own file
