+++ src/scripts.js
@@ -7,36 +7,37 @@ import { fetchData } from "./apiCalls";
 // An example of how you tell webpack to use an image (also need to link to it in the index.html)
 import "./images/turing-logo.png";
 import "./images/pexels-pixabay-258154.jpg";
+import domUpdates from './domUpdates'
 import { bookingsData } from "./data/bookingsData";
+import domUpdates from "./domUpdates";
 
 //  =========================================================================QUERY SELECTORS
-const homeBtn = document.querySelector("#homeDisplayLink");
-const dashboardBtn = document.querySelector("#dashboardLink");
-const reservationsBtn = document.querySelector("#reservationsLink");
+const navHomeBtn = document.querySelector("#navHomeBtn");
+const navDashboardBtn = document.querySelector("#navDashboardBtn");
+const navBookBtn = document.querySelector("#navBookBtn");
+
 const homeDisplay = document.querySelector("#homeDisplay");
 const dashboardDisplay = document.querySelector("#dashboardDisplay");
-const bookingDisplay = document.querySelector("#bookingDisplay");
+const bookDisplay = document.querySelector("#bookDisplay");
+
 const headDashboard = document.querySelector("#headDashboard");
 const bodyDashboard = document.querySelector("#bodyDashboard");
+const footerDashboard = document.querySelector("#footerDashboard");
+
 const dashboardContainer = document.querySelector("#dashboardContainer");
-const bookingContainer = document.querySelector("#bookingContainer");
-const bookingHistoryCard = document.querySelector("#bookingHistoryCard");
-const managerBookingHistory = document.querySelector("#managerBookingHistory");
+const bookContainer = document.querySelector("#bookContainer");
+const customerBookingHistory = document.querySelector("#customerBookingHistory");
+
 const bookingOptions = document.querySelector("#bookingOptions");
 const startDate = document.querySelector("#startDate");
 const roomTypes = document.querySelector("#roomTypes");
-const roomSearchBtn = document.querySelectorAll("#roomSearchBtn");
-const customerSearchInput = document.querySelector("#customerSearchInput");
-const customerSearchBtn = document.querySelector("#customerSearchBtn");
+const bookSearchBtn = document.querySelector("#bookSearchBtn");
+
 const searchMsg = document.querySelector("#searchMsg");
 const totalCost = document.querySelector("#totalCost");
 const headDashboardContainer = document.querySelector("#headDashboardContainer");
 const cards = document.querySelector("#cards");
-const validationMessage = document.querySelector("#validateMsg");
-const allBookings = document.querySelector("#allBookings");
-const pastBookings = document.querySelector("#pastBookings");
-const presentBookings = document.querySelector("#presentBookings");
-const futureBookings = document.querySelector("#futureBookings");
+
 const currentUsername = document.querySelector("#currentUsername");
 const currentPassword = document.querySelector("#currentPassword");
 const loginBtn = document.querySelector("#loginBtn");
@@ -54,7 +55,6 @@ const loadData = () => {
   const getAllCustomers = fetchData("http://localhost:3001/api/v1/customers", "GET");
   const getAllRooms = fetchData("http://localhost:3001/api/v1/rooms", "GET");
   const getbookings = fetchData("http://localhost:3001/api/v1/bookings", "GET");
-
   Promise.all([getAllCustomers, getAllRooms, getbookings])
     .then((data) => {
       allCustomerData = data[0].customers.map((customer) => new Customer(customer));
@@ -65,23 +65,15 @@ const loadData = () => {
     .catch((error) => console.log(error));
 };
 
-const getBookingAPI = () => {
+const getBookingFromAPI = () => {
   let url = "http://localhost:3001/api/v1/bookings";
   let methodType = "GET";
-
   Promise.all([fetchData(url, methodType)])
     .then((data) => {
       allBookingData = [];
       allBookingData = data[0].bookings.map((booking) => new Booking(booking));
       allBookingData.map((booking) => booking.setRoom(allRoomData));
-
       currentCustomer.addBookings(allBookingData);
-
-      displayManagerDashboard(
-        getAvailableRooms(getCurrentDate(), "all rooms"),
-        getBookedRooms(),
-        getTotalSales()
-      );
       displayDashboardCards(currentCustomer.bookings, false);
       displayAvailableBookings(getStartDate(), roomTypes.value);
     })
@@ -100,7 +92,7 @@ const addNewBooking = (roomNumber) => {
   Promise.all([postBookingAPI(roomNumber)])
     .then((data) => {
       console.log(data[0]);
-      getBookingAPI();
+      getBookingFromAPI();
     })
     .catch((error) => console.log(error));
 };
@@ -122,10 +114,11 @@ const deleteBooking = (bookingToBeDeleted) => {
 
 // =========================================================================EVENT LISTENERS
 window.addEventListener("load", loadData);
-dashboardBtn.addEventListener("click", () => loadDashboardDisplay());
-homeBtn.addEventListener("click", () => loadHomeDisplay());
+navHomeBtn.addEventListener("click", loadHomeDisplay);
+navDashboardBtn.addEventListener("click", loadDashboardDisplay);
+navBookBtn.addEventListener("click", () => loadBookView());
+
 bookingOptions.addEventListener("click", (event) => displayBookings(event));
-reservationsBtn.addEventListener("click", () => displayReservations());
 dashboardContainer.addEventListener("click", (event) => {
   let bookingId = event.target.id;
   if (!(getDateForBooking(bookingId) >= getCurrentDate())) {
@@ -135,183 +128,25 @@ dashboardContainer.addEventListener("click", (event) => {
     deleteBooking(bookingId);
   }
 });
-bookingContainer.addEventListener("click", (event) => {
+bookContainer.addEventListener("click", (event) => {
   let roomNumber = Number(event.target.id);
   if (roomNumber) {
     addNewBooking(roomNumber);
   }
 });
 
-roomSearchBtn.addEventListener("click", (event) => {
-  event.preventDefault();
-  displayAvailableBookings(getStartDate(), roomTypes.value);
-});
+  bookSearchBtn.addEventListener("click", (event) => {
+    event.preventDefault();
+    displayAvailableBookings(getStartDate(), roomTypes.value);
+  });
 
-customerSearchBtn.addEventListener("click", (event) => {
-  event.preventDefault();
-  currentCustomer = getCustomerData(customerSearchInput.value);
-  if (!currentCustomer) {
-    displaySearchMsg("incorrect or invalid customer name");
-    currentCustomer = "";
-  }
-  loadDashboardDisplay();
-  customerSearchInput.value = "";
-});
-
-//  ========================================================================= DOM
-const hideElemet = (element) => {
-    element.classList.add("hidden");
-  };
-  const showElement = (element) => {
-    element.classList.remove("hidden");
-  };
-  
-  const clearDisplay = (view) => {
-    view.innerHTML = "";
-  };
-  
-  const displaySearchMsg = (message) => {
-    searchMsg.innerText = message;
-  };
-  
-  const displayManagerSearchMsg = (message) => {
-    dashboardContainer.innherHTML = `<h2 class='my-4 text-center md'>${message}</h2>`;
-  };
-  
-  const displayDashboardCards = (bookings, isHidden = true) => {
-    clearDisplay(dashboardContainer);
-    if (!bookings.length) {
-      dashboardContainer.innerHTML = `<h2>there are no bookings available</h2>`;
-    } else {
-      displayBookedCards(bookings, isHidden);
-    }
-    displayBookingCost(bookings);
-  };
-  
-  const displayBookedCards = (bookings, isHidden) => {
-    let hidBtn = "";
-    if (isHidden) {
-      hidBtn = "hidden";
-    }
-    bookings.forEach((booking) => {
-      dashboardContainer.innerHTML += `
-          <div class="card flex">
-              <summary>
-                  <h3>room ${booking.roomDetails.number}</h3>
-                  <h3>room ${booking.roomDetails.roomType}</h3>
-                  <ul>
-                      <li class="sm"><span>beds:</span> ${
-                        booking.roomDetails.numBeds
-                      }</li>
-                      <li class="sm"><span>bed size:</span> ${
-                        booking.roomDetails.bedSize
-                      }</li>
-                      <li class="sm"><span>bidet:</span> ${
-                        booking.roomDetails.bidet
-                      }</li>
-                      <li class="sm"><span>date:</span> ${booking.date}</li>
-                      <li class="sm"><span>cost per night:</span> ${booking.roomDetails.costPerNight.toFixed(
-                        2
-                      )}</li>
-                      <li class="sm"><span>beds:</span> ${
-                        booking.roomDetails.numBeds
-                      }</li>
-                  </ul>
-              </summary>
-              <button class="${hidBtn}" id="${booking.id}>delete</button
-          </div>`;
-    });
-  };
-  
-  const displayBookingCost = (bookings) => {
-    const bookingTotal = bookings.reduce((acc, booking) => {
-      acc += booking.roomDetails.costPerNight;
-      return acc;
-    }, 0);
-    return (totalCost.innerText = `${bookingTotal.toFixed(2)}`);
-  };
-  
-  const displayBookingCards = (startDate, availableRooms) => {
-    clearDisplay(bookingContainer);
-    if (!availableRooms.length) {
-      boookingContainer.innerHTML = `
-          <h1>no available bookings</h1>`;
-    }
-    displayAvailableRooms(bookingContainer, startDate, availableRooms);
-  };
-  
-  const displayAvailableRooms = (container, startDate, availableRooms) => {
-    availableRooms.forEach((room) => {
-      container.innerHTML += `
-          <div class="flex card">
-              <summary>
-                  <h3>room ${room.number}</h3>
-                  <h3>${room.roomType}</h3>
-                  <ul>
-                  <li class="sm"><span>beds:</span> ${room.numBeds}</li>
-                  <li class="sm"><span>bed size:</span> ${room.bedSize}</li>
-                  <li class="sm"><span>bidet:</span> ${room.bidet}</li>
-                  <li class="sm"><span>date:</span> ${startDate}</li>
-                  <li class="sm"><span>cost per night:</span> ${room.costPerNight.toFixed(
-                    2
-                  )}</li>
-              </ul>
-              </summary>
-              <button class="btn" id="${room.number}">reserve</button>
-          </div>`;
-    });
-  };
-  
-  const displayDashboardHeader = (currentCustomer) => {
-    const displayName = `${currentCustomer.name}`;
-    const welcomeMsg = "Welcome to Hotel Overlook!";
-    displayDashHead(displayName, welcomeMsg);
-  };
-  
-  const displayDashHead = (displayName, welcomeMsg) => {
-    headDashboardContainer.innerHTML += `
-      <article>
-          <h2 class="lg">${displayName}</h2>
-          <p>Date: ${getCurrentDate()}</p>
-          <p>${welcomeMsg}</p>
-      </article>`;
-  };
-  
-  const displayManagerDashboard = (roomsAvailable, bookedRooms, salesTotal) => {
-    let availableRoomCount = roomsAvailable.length;
-    let occupiedRoomPercent = Math.round((bookedRooms.length / 25) * 100);
-    cards.innerHTML = `
-          <h2 class="md my-3 text-center">Info for ${getCurrentDate()}</h2>
-          <h3 class="my-2">Available rooms</h2>
-          <p class="sm">${availableRoomCount}</p>
-          <h3 class="my-2">Occupied rooms</h2>
-          <p class="sm">${occupiedRoomPercent}</p>
-          <h3 class="my-2">Total sales</h2>
-          <p class="sm">${salesTotal}</p>
-      `;
-  };
-  
-  const getCurrentDate = () => {
-    let today = new Date();
-    let dd = today.getDate();
-    let mm = today.getMonth() + 1;
-    let yyyy = today.getFullYear();
-    if (Number(dd) < 10) {
-      dd = `0${dd}`;
-    } else if (Number(mm) < 10) {
-      mm = `0${mm}`;
-    }
-    return `${yyyy}/${mm}/${dd}`;
-  };
-  
-  
 //  =========================================================================FUNCTIONS
 const hideElements = () => {
   hideElement(homeDisplay);
   hideElement(dashboardDisplay);
-  hideElement(bookingHistoryCard);
+  hideElement(customerBookingHistory);
   hideElement(managerBookingHistory);
-  hideElement(bookingDisplay);
+  hideElement(bookDisplay);
   hideElement(loginDisplay);
   hideElement(logoutDisplay);
   hideElement(headDashboard);
@@ -339,55 +174,36 @@ const loadHomeDisplay = () => {
 };
 
 const loadDashboardDisplay = () => {
-  if (!currentCustomer && !currentManager) {
+  if (!currentCustomer) {
     loginError("Please login");
-  } else if (currentManager !== "") {
-    hideElements();
-    showElement(dashboardDisplay);
-    showElement(managerBookingHistory);
-    showElement(headDashboard);
-    showElement(bodyDashboard);
-    displayManagerDashboard(
-      getAvailableRooms(getCurrentDate(), "all rooms"),
-      getBookedRooms(),
-      getTotalSales()
-    );
-    displayDashboardHeader(currentManager);
-
-    if (currentCustomer !== "") {
-      displaySearchMsg("");
-      displayDashboardCards(currentCustomer.bookings, false);
-    } else {
-      displayManagerSearchMsg("");
-    }
   } else {
     hideElements();
     showElement(dashboardDisplay);
     showElement(headDashboard);
     showElement(bodyDashboard);
-    showElement(bookingHistoryCard);
+    showElement(customerBookingHistory);
     displayDashboardHeader(currentCustomer);
     displayDashboardCards(currentCustomer.bookings);
   }
 };
 
 const displayReservations = () => {
-  if (!currentCustomer && !currentManager) {
+  if (!currentCustomer) {
     loginError("please log in");
   } else if (!currentCustomer) {
     loginError("invalid user");
   } else {
     hideElements();
-    showElement(bookingDisplay);
+    showElement(bookDisplay);
     showElement(headDashboard);
     showElement(bodyDashboard);
-    resetBookingDisplay();
+    resetbookDisplay();
     displayAvailableBookings(getCurrentDate(), roomTypes.value);
     displayBookingHeader();
   }
 };
 
-const resetBookingDisplay = () => {
+const resetbookDisplay = () => {
   roomTypes.value = "all rooms";
   startDate.value = getCurrentDate().replaceAll("/", "-");
 };
